+++ huaweicloud-sdk-java-dis-iface/src/main/java/com/huaweicloud/dis/iface/app/response/AppEntry.java
@@ -13,15 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.huaweicloud.dis.iface.app.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
-/**
- * Created by z00382129 on 2017/11/7.
- */
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class AppEntry {