+++ huaweicloud-sdk-java-dis-iface/src/main/java/com/huaweicloud/dis/iface/coordinator/ICoordinatorService.java
@@ -8,9 +8,6 @@ import com.huaweicloud.dis.iface.coordinator.response.HeartbeatResponse;
 import com.huaweicloud.dis.iface.coordinator.response.JoinGroupResponse;
 import com.huaweicloud.dis.iface.coordinator.response.SyncGroupResponse;
 
-/**
- * Created by z00382129 on 2017/10/17.
- */
 public interface ICoordinatorService
 {
     HeartbeatResponse handleHeartbeatRequest(HeartbeatRequest heartbeatRequest);