+++ huaweicloud-sdk-java-dis-iface/src/main/java/com/huaweicloud/dis/iface/stream/response/PartitionResult.java
@@ -20,9 +20,6 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-/**
- * Created by kienluu on 7/19/17.
- */
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class PartitionResult