+++ huaweicloud-sdk-java-dis/src/main/java/com/huaweicloud/dis/http/AbstractDISClient.java
@@ -306,7 +306,7 @@ public class AbstractDISClient {
     
     private void setParameters(Request<HttpRequest> request, Object requestContent)
     {
-        if (request.getHttpMethod().equals(HttpMethodName.GET))
+        if (request.getHttpMethod().equals(HttpMethodName.GET) || request.getHttpMethod().equals(HttpMethodName.DELETE))
         {
             if (requestContent != null)
             {
@@ -385,10 +385,10 @@ public class AbstractDISClient {
 	private <T> Future<T> doRequestAsync(Request<HttpRequest> request, Object requestContent, String ak, String sk,
 			String region, Class<T> returnType, AsyncHandler<T> callback) {
 		String uri = buildURI(request);
-		
+
 		request.getHeaders().remove(SignerConstants.AUTHORIZATION);
-		request = SignUtil.sign(request, ak, sk, region);
-		
+        request = SignUtil.sign(request, ak, sk, region, disConfig);
+
 		ConnectRetryFuture<T> connectRetryFuture = new ConnectRetryFuture<T>(request, ak, sk, requestContent, callback, uri, returnType);
 		
 		ConnectRetryCallback<T> connectRetryCallback = null;
@@ -489,7 +489,7 @@ public class AbstractDISClient {
             	int tmpRetryIndex = retryCount.incrementAndGet();
                 
                 request.getHeaders().remove(SignerConstants.AUTHORIZATION);
-                request = SignUtil.sign(request, ak, sk, region);
+                request = SignUtil.sign(request, ak, sk, region,disConfig);
                 
                 ConnectRetryCallback<T> connectRetryCallback = null;
                 if(callback != null){
@@ -497,7 +497,7 @@ public class AbstractDISClient {
                 }
                 
                 LOG.warn("connect or system error retry [{}] [{}] [{}]", this.hashCode(), retryIndex, errorMsg);
-                Future<T> restFuture = RestClientAsync.getInstance(disConfig).exchangeAsync(uri, 
+                Future<T> restFuture = RestClientAsync.getInstance(disConfig).exchangeAsync(uri,
                 		request.getHttpMethod(), request.getHeaders(), requestContent, returnType, connectRetryCallback);
                 
                 this.setInnerFuture(restFuture);
@@ -609,7 +609,7 @@ public class AbstractDISClient {
             {
                 request.getHeaders().remove(SignerConstants.AUTHORIZATION);
                 // 每次重传需要重新签名
-                request = SignUtil.sign(request, ak, sk, region);
+                request = SignUtil.sign(request, ak, sk, region,disConfig);
                 return RestClient.getInstance(disConfig).exchange(uri, 
                 		request.getHttpMethod(), request.getHeaders(), requestContent, returnType);
             }
@@ -643,7 +643,7 @@ public class AbstractDISClient {
 	 *
 	 * @param t
 	 *            throwable exception
-	 * @param request
+	 * @param request HttpReuest
      * @return {@code true} retriable {@code false} not retriable
 	 */
     protected boolean isRetriableSendException(Throwable t, Request<HttpRequest> request)