+++ huaweicloud-sdk-java-dis/src/main/java/com/huaweicloud/dis/http/ProxyAuthenticationMethod.java
@@ -14,24 +14,15 @@
  * limitations under the License.
  */
 
-package com.huaweicloud.dis.iface.app.request;
+package com.huaweicloud.dis.http;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class CreateAppRequest {
-
-    @JsonProperty("app_name")
-    private String appName;
-
-    public String getAppName() {
-        return appName;
-    }
-
-    public void setAppName(String appName) {
-        this.appName = appName;
-    }
+/**
+ * Methods by which the SDK can authenticate against an HTTP proxy. These can be configured on a client using
+ */
+public enum ProxyAuthenticationMethod {
+    SPNEGO,
+    KERBEROS,
+    NTLM,
+    DIGEST,
+    BASIC
 }