+++ CHANGELOG.md
@@ -6,3 +6,12 @@
   * add delete checkpoint
   * add describe consumer state
   * support tags in stream apis
+
+# 1.3.2
+
+- Features
+  * Support configuring proxy
+  * Support caching failed records locally
+
+- Bugfixs
+  * update jackson-core version to 2.8.11